soc_core/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "soc-core"
version = "1.0.0"
description = "Project Synapse SOC Factory Core"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.1.3",
    "requests>=2.33.1",
    "python-dotenv>=1.2.2",
    "PyYAML>=6.0.3",
    "chromadb>=1.5.5",
    "pydantic>=2.12.5",
    "google-genai>=1.70.0",
    "google-generativeai>=0.8.6",
    "rich>=14.3.3",
    "tqdm>=4.67.3"
]

[tool.hatch.build.targets.wheel]
packages = ["core", "soc", "api", "scripts"]
